automation/tests/integration/requirements.txt
requests==2.31.0
pyyaml==6.0.1
typing-extensions==4.9.0 